Updated: General feature enhancements and code standards for ezmessage module

+++ kernel/ezmessage/admin/templates/standard/messagesend.php
@@ -1,6 +1,6 @@
 <?php
 // 
-// $Id: messagelist.php 6316 2001-07-30 13:22:37Z bf $
+// $Id: messagelist.php,v 1.4 2001/07/30 13:22:37 bf Exp $
 //
 // Created on: <05-Jun-2001 16:42:09 bf>
 //
@@ -23,16 +23,16 @@
 // Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, US
 //
 
-// include_once( "classes/ezlocale.php" );
-// include_once( "classes/ezhttptool.php" );
-// include_once( "classes/eztemplate.php" );
-// include_once( "classes/INIFile.php" );
+include_once( "classes/ezlocale.php" );
+include_once( "classes/ezhttptool.php" );
+include_once( "classes/eztemplate.php" );
+include_once( "classes/INIFile.php" );
 
-// include_once( "ezmessage/classes/ezmessage.php" );
+include_once( "ezmessage/classes/ezmessage.php" );
 
 
-$t = new eZTemplate( "kernel/ezmessage/admin/" . $ini->read_var( "eZMessageMain", "AdminTemplateDir" ),
-                     "kernel/ezmessage/admin/intl", $Language, "messagelist.php" );
+$t = new eZTemplate( "ezmessage/admin/" . $ini->read_var( "eZMessageMain", "AdminTemplateDir" ),
+                     "ezmessage/admin/intl", $Language, "messagelist.php" );
 
 $locale = new eZLocale( $Language ); 
 
@@ -44,26 +44,10 @@ if ( isSet( $Delete ) && isSet( $DelMessage ) && count( $DelMessage ) > 0 )
 	{
 		$message = new eZMessage( $value );
 		$toUser = $message->toUser();
-		$fromUser = $message->fromUser();
 		$user =& eZUser::currentUser();
-		
-		if ( $toUser->id() == $user->id() )
+		if ( $toUser->id() == $user->id() && $message->isRead() == true )
 		{
-			
-			$messageDefinition = new eZMessageDefinition();
-			$definitionArray = $messageDefinition->getMessageID( $value );
-
-			foreach ( $definitionArray as $definition )
-			{
-				$to_user = $definition->toUserID();
-				
-				if ( $to_user->id() == $user->id() )
-					$definition->delete();
-			}
-			//print_r($definitionArray);
-			//echo count($definitionArray);
-			if ( count( $definitionArray ) == 1 )
-				$message->delete();
+			$message->delete();
 		}
 	}
 }
@@ -105,8 +89,6 @@ foreach ( $messageArray as $key => $message )
     {
         $t->set_var( "message_disabled", "" );
         $t->set_var( "message_unread", "" );
-        $timeread = $message->TimeRead();
-        $t->set_var( "time_read", $locale->format( $timeread ) );
         $t->parse( "message_read", "message_read_tpl" );
     }
     else
@@ -124,6 +106,8 @@ if ( count( $messageArray ) > 0 )
 else
     $t->set_var( "message_list", "" );
 
+
 $t->pparse( "output", "message_page_tpl" );
 
 ?>
+
